Move marker interfaces back to spring
