+++ examples/spring/src/main/kotlin/com/expediagroup/graphql/examples/mutation/WidgetMutation.kt
@@ -18,7 +18,7 @@ package com.expediagroup.graphql.examples.mutation
 
 import com.expediagroup.graphql.annotations.GraphQLDescription
 import com.expediagroup.graphql.examples.model.Widget
-import com.expediagroup.graphql.types.operations.Mutation
+import com.expediagroup.graphql.spring.operations.Mutation
 import org.springframework.stereotype.Component
 
 /**
