+++ examples/spring/src/main/kotlin/com/expediagroup/graphql/examples/query/OrganizationQuery.kt
@@ -3,7 +3,7 @@ package com.expediagroup.graphql.examples.query
 import com.expediagroup.graphql.annotations.GraphQLDescription
 import com.expediagroup.graphql.examples.model.Company
 import com.expediagroup.graphql.examples.model.Employee
-import com.expediagroup.graphql.types.operations.Query
+import com.expediagroup.graphql.spring.operations.Query
 import graphql.schema.DataFetcher
 import graphql.schema.DataFetchingEnvironment
 import org.springframework.beans.factory.BeanFactory
