+++ graphql-kotlin-spring-server/src/test/kotlin/com/expediagroup/graphql/spring/execution/DataFetcherIT.kt
@@ -16,7 +16,7 @@
 package com.expediagroup.graphql.spring.execution
 
 import com.expediagroup.graphql.hooks.SchemaGeneratorHooks
-import com.expediagroup.graphql.types.operations.Query
+import com.expediagroup.graphql.spring.operations.Query
 import com.expediagroup.graphql.types.GraphQLRequest
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
 import graphql.language.StringValue
