+++ graphql-kotlin-spring-server/src/test/kotlin/com/expediagroup/graphql/spring/execution/IntrospectionIT.kt
@@ -15,7 +15,7 @@
  */
 package com.expediagroup.graphql.spring.execution
 
-import com.expediagroup.graphql.types.operations.Query
+import com.expediagroup.graphql.spring.operations.Query
 import com.expediagroup.graphql.types.GraphQLRequest
 import graphql.introspection.IntrospectionQuery
 import org.junit.jupiter.api.Test
