+++ graphql-kotlin-spring-server/src/test/kotlin/com/expediagroup/graphql/spring/instrumentation/InstrumentationIT.kt
@@ -1,7 +1,7 @@
 package com.expediagroup.graphql.spring.instrumentation
 
 import com.expediagroup.graphql.spring.DEFAULT_INSTRUMENTATION_ORDER
-import com.expediagroup.graphql.types.operations.Query
+import com.expediagroup.graphql.spring.operations.Query
 import com.expediagroup.graphql.types.GraphQLRequest
 import graphql.ExecutionResult
 import graphql.ExecutionResultImpl
